+++ src/Ctrl/UsuarioCtrl.js
@@ -82,4 +82,22 @@ export const getUsuarioByID = async (id) => {
   }
 };
 
-export default { verificarUsuario, registrarUsuario, getUsuarioByEmail, updateUsuario , getUsuarioByID};
+export const getNombresApellidosById = async (userId) => {
+    try {
+        const { data, error } = await supabase.from("usuario").select("nombres, apellidos").eq("idusuario", userId);
+
+        if (error) {
+            console.error("Error obteniendo el usuario:", error.message);
+            throw new Error("No se pudo obtener el usuario: " + error.message);
+        }
+
+        if (data.length > 0) { return `${data[0].nombres} ${data[0].apellidos}`;
+        } else { return "Usuario desconocido"; }
+
+    } catch (error) {
+        console.error("Error interno:", error.message);
+        throw new Error("Ocurrió un error al obtener la información del usuario: " + error.message);
+    }
+};
+
+export default { verificarUsuario, registrarUsuario, getUsuarioByEmail, updateUsuario , getUsuarioByID, getNombresApellidosById};
