+++ src/routes/CalendarioEventos.jsx
@@ -4,7 +4,7 @@ import React, { useState, useEffect } from 'react';
 import "react-big-calendar/lib/css/react-big-calendar.css";
 import { CiCalendarDate } from "react-icons/ci";
 import "dayjs/locale/es";
-import { getAllEventIds, getEventInfo } from "../Ctrl/EventosCtrl";
+import { getAllEventIds, getEventById } from "../Ctrl/EventosCtrl";
 import { getUsuarioByID } from "../Ctrl/UsuarioCtrl";
 import { useNavigate } from 'react-router-dom';
 import './CalendarioEventosStyles.css';
@@ -27,7 +27,7 @@ const Calendario = () => {
 
             const eventDetails = await Promise.all(
                 eventIds.map(async (id) => {
-                    const event = await getEventInfo(id);
+                    const event = await getEventById(id);
                     if (!event || (event.estado !== "Confirmado" && event.estado !== "Pendiente")) {
                         console.warn(`Evento con ID ${id} tiene un estado no válido.`);
                         return null;
